+++ examples/ai-core/src/stream-text/google-grounding.ts
@@ -4,11 +4,9 @@ import 'dotenv/config';
 
 async function main() {
   const result = streamText({
-    model: google('gemini-2.0-flash-exp'),
-    providerOptions: {
-      google: {
-        useSearchGrounding: true,
-      },
+    model: google('gemini-2.5-flash'),
+    tools: {
+      google_search: google.tools.googleSearch({}),
     },
     prompt: 'List the top 5 San Francisco news from the past week.',
   });