+++ packages/ai/src/ui-message-stream/read-ui-message-stream.test.ts
@@ -1,0 +1,99 @@
+import {
+  convertArrayToReadableStream,
+  convertAsyncIterableToArray,
+} from '@ai-sdk/provider-utils/test';
+import { UIMessageChunk } from './ui-message-chunks';
+import { readUIMessageStream } from './read-ui-message-stream';
+
+function createUIMessageStream(parts: UIMessageChunk[]) {
+  return convertArrayToReadableStream(parts);
+}
+
+describe('readUIMessageStream', () => {
+  it('should return a ui message object stream for a basic input stream', async () => {
+    const stream = createUIMessageStream([
+      { type: 'start', messageId: 'msg-123' },
+      { type: 'start-step' },
+      { type: 'text-start', id: 'text-1' },
+      { type: 'text-delta', id: 'text-1', delta: 'Hello, ' },
+      { type: 'text-delta', id: 'text-1', delta: 'world!' },
+      { type: 'text-end', id: 'text-1' },
+      { type: 'finish-step' },
+      { type: 'finish' },
+    ]);
+
+    const uiMessages = readUIMessageStream({ stream });
+
+    expect(await convertAsyncIterableToArray(uiMessages))
+      .toMatchInlineSnapshot(`
+      [
+        {
+          "id": "msg-123",
+          "metadata": undefined,
+          "parts": [],
+          "role": "assistant",
+        },
+        {
+          "id": "msg-123",
+          "metadata": undefined,
+          "parts": [
+            {
+              "type": "step-start",
+            },
+            {
+              "state": "streaming",
+              "text": "",
+              "type": "text",
+            },
+          ],
+          "role": "assistant",
+        },
+        {
+          "id": "msg-123",
+          "metadata": undefined,
+          "parts": [
+            {
+              "type": "step-start",
+            },
+            {
+              "state": "streaming",
+              "text": "Hello, ",
+              "type": "text",
+            },
+          ],
+          "role": "assistant",
+        },
+        {
+          "id": "msg-123",
+          "metadata": undefined,
+          "parts": [
+            {
+              "type": "step-start",
+            },
+            {
+              "state": "streaming",
+              "text": "Hello, world!",
+              "type": "text",
+            },
+          ],
+          "role": "assistant",
+        },
+        {
+          "id": "msg-123",
+          "metadata": undefined,
+          "parts": [
+            {
+              "type": "step-start",
+            },
+            {
+              "state": "done",
+              "text": "Hello, world!",
+              "type": "text",
+            },
+          ],
+          "role": "assistant",
+        },
+      ]
+    `);
+  });
+});