+++ packages/codemod/src/test/remove-experimental-useassistant.test.ts
@@ -1,5 +1,5 @@
 import { describe, it } from 'vitest';
-import transformer from '../codemods/remove-experimental-useassistant';
+import transformer from '../codemods/v4/remove-experimental-useassistant';
 import { testTransform } from './test-utils';
 
 describe('remove-experimental-useassistant', () => {