+++ packages/deepinfra/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @ai-sdk/deepinfra
 
+## 1.0.0-beta.5
+
+### Patch Changes
+
+- Updated dependencies [05d2819]
+  - @ai-sdk/provider-utils@3.0.0-beta.3
+  - @ai-sdk/openai-compatible@1.0.0-beta.5
+
+## 1.0.0-beta.4
+
+### Patch Changes
+
+- Updated dependencies [1b101e1]
+  - @ai-sdk/openai-compatible@1.0.0-beta.4
+
 ## 1.0.0-beta.3
 
 ### Patch Changes