+++ packages/mistral/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @ai-sdk/mistral
 
+## 2.0.0-beta.4
+
+### Patch Changes
+
+- Updated dependencies [05d2819]
+  - @ai-sdk/provider-utils@3.0.0-beta.3
+
+## 2.0.0-beta.3
+
+### Patch Changes
+
+- c3b7069: add Mistral Medium 3 model support
+
 ## 2.0.0-beta.2
 
 ### Patch Changes