+++ packages/perplexity/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @ai-sdk/perplexity
 
+## 2.0.0-beta.4
+
+### Patch Changes
+
+- Updated dependencies [05d2819]
+  - @ai-sdk/provider-utils@3.0.0-beta.3
+
+## 2.0.0-beta.3
+
+### Major Changes
+
+- 2a9732b: feat(provider/perplexity): support image input
+
 ## 2.0.0-beta.2
 
 ### Patch Changes